requirements.txt
streamlit
pandas
numpy
sentence-transformers
faiss-cpu
openai